+++ domains/java/java-data-structure/java-comparator/Player.java
@@ -1,0 +1,9 @@
+class Player {
+  String name;
+  int score;
+
+  Player(String name, int score) {
+    this.name = name;
+    this.score = score;
+  }
+}
