+++ domains/java/java-data-structure/java-comparator/Solution.java
@@ -1,5 +1,4 @@
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.Scanner;
 
 class Solution {
@@ -16,29 +15,5 @@ class Solution {
     for (int i = 0; i < player.length; i++) {
       System.out.printf("%s %s\n", player[i].name, player[i].score);
     }
-  }
-}
-
-class Checker implements Comparator {
-  public int compare(Object o1, Object o2) {
-    Player p1 = (Player) o1;
-    Player p2 = (Player) o2;
-    if (p1.score == p2.score) {
-      return p1.name.compareTo(p2.name);
-    } else if (p1.score < p2.score) {
-      return 1;
-    } else {
-      return -1;
-    }
-  }
-}
-
-class Player {
-  String name;
-  int score;
-
-  Player(String name, int score) {
-    this.name = name;
-    this.score = score;
   }
 }
